Make Civetweb use atomic operations for STOP_FLAG
On our CV machines we've seen mg_stop() of Civetweb hang due to it not
seeing the state of the stop flag change to value 2. To fix this ensure
we're using atomic read/write functions. Which is enable by setting the
defs STOP_FLAG_NEEDS_LOCK and to enable the functions
STOP_FLAG_NEEDS_LOCK needs set USE_SERVER_STATS.